url_shortener_backend/requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
Pygments==2.6.1
pytz==2020.1
Rx==1.6.1
short-url==1.2.2
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
